+++ cmake/config.h.in
@@ -19,9 +19,6 @@
 /* Define to 1 to use the gflags package for command-line parsing. */
 #cmakedefine HAVE_GFLAGS 1
 
-/* Define to 1 if you have Google Test. */
-#cmakedefine HAVE_GTEST 1
-
 /* Define to 1 if you have the `lzo2' library (-llzo2). */
 #cmakedefine HAVE_LIBLZO2 1
